Add message bus to abstract batch system
