+++ src/toil/batchSystems/abstractBatchSystem.py
@@ -131,7 +131,6 @@ class AbstractBatchSystem(ABC):
         Currently the only message a batch system may send is
         JobAnnotationMessage.
         """
-        # By default, do nothing.
         pass
 
     @abstractmethod
@@ -311,6 +310,7 @@ class BatchSystemSupport(AbstractBatchSystem):
                 workflow_id=config.workflowID,
                 clean_work_dir=config.cleanWorkDir,
             )
+        self._outbox: Optional[MessageOutbox] = None
 
     def checkResourceRequest(self, memory: int, cores: float, disk: int, job_name: str = '', detail: str = '') -> None:
         """
@@ -377,6 +377,18 @@ class BatchSystemSupport(AbstractBatchSystem):
             except KeyError:
                 raise RuntimeError(f"{name} does not exist in current environment")
         self.environment[name] = value
+
+    def set_message_bus(self, message_bus: MessageBus) -> None:
+        """
+        Give the batch system an opportunity to connect directly to the message
+        bus, so that it can send informational messages about the jobs it is
+        running to other Toil components.
+
+        Currently the only message a batch system may send is
+        JobAnnotationMessage.
+        """
+        # We do in fact send messages to the message bus.
+        self._outbox = message_bus.outbox()
 
     def formatStdOutErrPath(self, toil_job_id: int, cluster_job_id: str, std: str) -> str:
         """
